requirements.txt
instaloader==4.5.3
InstagramAPI==1.0.2
instabot==0.117.0
psycopg2-binary==2.8.6
python-decouple==3.3
requests==2.11.1
moviepy==0.2.3.2
requests-toolbelt==0.7.0
python-dateutil==2.8.1
requests==2.24.0